Adding tests for C++ and python (#40)
* Adding tests for C++ and python

* fix the test cases

* control the flag correctly for GPUFL variables

* trying to fix windows issue

* fix windows action

* fix window build

* Fix the build issue

* remove discovery mode

* fix build

* Adding PRE_TEST

* copy CUDA runtime DLLs

* fix the windows build

* fix the windows build

* Skip automatic test for windows for now

* Disable unit test for windows for now

+++ include/gpufl/core/stack_trace.cpp
@@ -30,7 +30,7 @@
                 symbol->SizeOfStruct = sizeof(SYMBOL_INFO);
 
                 bool first = true;
-                for (int i = frames - 1; i >= skipFrames; --i) {
+                for (int i = static_cast<int>(frames) - 1; i >= skipFrames; --i) {
                     if (SymFromAddr(process, (DWORD64)(stack[i]), 0, symbol)) {
                         std::string name = symbol->Name;
